Fix: Updated component structuring and maintain common utils.
