+++ .gitignore
@@ -9,6 +9,7 @@
 
 # misc
 .DS_Store
+.vscode
 .env.local
 .env.development.local
 .env.test.local
