+++ src/utils.js
@@ -75,3 +75,11 @@ export const genSubNavScrollPos = (data) => {
     stopPos: scrollPos
   }
 };
+
+export const genRandomKey = () => (
+  Math.random().toString(36).slice(2)
+)
+
+export const isMenuLight = pathname => (
+  pathname === '/get-involved/'
+)
