--- src/v1.1/utils.js
@@ -1,7 +1,0 @@
-export const genRandomKey = () => (
-  Math.random().toString(36).slice(2)
-)
-
-export const isMenuLight = pathname => (
-  pathname === '/get-involved/'
-)
